ref: diagutils and id attribute
