+++ internal/provider/resource_organization_repository.go
@@ -2,10 +2,7 @@ package provider
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"net/http"
-	"strings"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
@@ -14,8 +11,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/jianyuan/go-sentry/v2/sentry"
+	"github.com/jianyuan/terraform-provider-sentry/internal/diagutils"
 )
 
 var _ resource.Resource = &OrganizationRepositoryResource{}
@@ -39,13 +36,7 @@ func (r *OrganizationRepositoryResource) Schema(ctx context.Context, req resourc
 		MarkdownDescription: "Organization Repository resource. This resource manages Sentry's source code management integrations.",
 
 		Attributes: map[string]schema.Attribute{
-			"id": schema.StringAttribute{
-				MarkdownDescription: "The ID of this resource.",
-				Computed:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.UseStateForUnknown(),
-				},
-			},
+			"id": ResourceIdAttribute(),
 			"organization": schema.StringAttribute{
 				MarkdownDescription: "The slug of the organization the resource belongs to.",
 				Required:            true,
@@ -99,12 +90,12 @@ func (r *OrganizationRepositoryResource) Create(ctx context.Context, req resourc
 		},
 	)
 	if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Create error: %s", err))
+		diagutils.AddClientError(resp.Diagnostics, "create", err)
 		return
 	}
 
 	if err := data.Fill(data.Organization.ValueString(), *repo); err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Fill error: %s", err))
+		diagutils.AddFillError(resp.Diagnostics, err)
 		return
 	}
 
@@ -128,7 +119,7 @@ out:
 	for {
 		repos, apiResp, err := r.client.OrganizationRepositories.List(ctx, data.Organization.ValueString(), params)
 		if err != nil {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Read error: %s", err))
+			diagutils.AddClientError(resp.Diagnostics, "read", err)
 			return
 		}
 
@@ -146,12 +137,12 @@ out:
 	}
 
 	if foundRepo == nil {
-		resp.Diagnostics.AddError("Not Found", "No matching organization repository found")
+		diagutils.AddNotFoundError(resp.Diagnostics, "organization repository")
 		return
 	}
 
 	if err := data.Fill(data.Organization.ValueString(), *foundRepo); err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Fill error: %s", err))
+		diagutils.AddFillError(resp.Diagnostics, err)
 		return
 	}
 
@@ -159,7 +150,7 @@ out:
 }
 
 func (r *OrganizationRepositoryResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	resp.Diagnostics.AddError("Not Supported", "Update is not supported for this resource")
+	diagutils.AddNotSupportedError(resp.Diagnostics, "update")
 }
 
 func (r *OrganizationRepositoryResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
@@ -179,7 +170,7 @@ func (r *OrganizationRepositoryResource) Delete(ctx context.Context, req resourc
 		return
 	}
 	if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Delete error: %s", err))
+		diagutils.AddClientError(resp.Diagnostics, "delete", err)
 		return
 	}
 }
@@ -187,7 +178,7 @@ func (r *OrganizationRepositoryResource) Delete(ctx context.Context, req resourc
 func (r *OrganizationRepositoryResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	organization, integrationType, integrationId, id, err := splitFourPartID(req.ID, "organization", "integration_type", "integration_id", "id")
 	if err != nil {
-		resp.Diagnostics.AddError("Import Error", fmt.Sprintf("Invalid ID: %s", err))
+		diagutils.AddImportError(resp.Diagnostics, err)
 		return
 	}
 
@@ -203,31 +194,4 @@ func (r *OrganizationRepositoryResource) ImportState(ctx context.Context, req re
 	resp.Diagnostics.Append(resp.State.SetAttribute(
 		ctx, path.Root("id"), id,
 	)...)
-}
-
-type OrganizationRepositoryModel struct {
-	Id              types.String `tfsdk:"id"`
-	Organization    types.String `tfsdk:"organization"`
-	IntegrationType types.String `tfsdk:"integration_type"`
-	IntegrationId   types.String `tfsdk:"integration_id"`
-	Identifier      types.String `tfsdk:"identifier"`
-}
-
-func (m *OrganizationRepositoryModel) Fill(organization string, repo sentry.OrganizationRepository) error {
-	m.Id = types.StringValue(repo.ID)
-	m.Organization = types.StringValue(organization)
-	m.IntegrationType = types.StringValue(strings.TrimPrefix(repo.Provider.ID, "integrations:"))
-	m.IntegrationId = types.StringValue(repo.IntegrationId)
-
-	var identifierStr string
-	var identifierNum json.Number
-	if err := json.Unmarshal(repo.ExternalSlug, &identifierStr); err == nil {
-		m.Identifier = types.StringValue(identifierStr)
-	} else if err := json.Unmarshal(repo.ExternalSlug, &identifierNum); err == nil {
-		m.Identifier = types.StringValue(identifierNum.String())
-	} else {
-		return fmt.Errorf("failed to unmarshal identifier: %s", repo.ExternalSlug)
-	}
-
-	return nil
 }
